+++ test/scope/ScopeModel.spec.js
@@ -9,6 +9,16 @@ import {
 import { createScopeCommandMock } from './_mocks/scopeCommandMocks';
 import { THEME } from '../../src/assets/scripts/client/constants/themes';
 
+let sandbox; // using the sinon sandbox ensures stubs are restored after each test
+
+ava.beforeEach(() => {
+    sandbox = sinon.sandbox.create();
+});
+
+ava.afterEach(() => {
+    sandbox.restore();
+});
+
 ava('does not throw when instantiated with no parameters', (t) => {
     t.notThrows(() => new ScopeModel());
 });
@@ -141,15 +151,50 @@ ava('.setScratchpad() sets RadarTargetModel._scratchPadText to the specified str
     t.true(radarTargetModelSetScratchPadSpy.calledWithExactly(newScratchPadText));
 });
 
-ava('.toggleHalo() sets RadarTargetModel._hasHalo to the opposite of its current state', (t) => {
-    const model = new ScopeModel();
+ava('.setHalo() returns error when requested halo size is invalid', (t) => {
+    const scopeModel = new ScopeModel();
     const radarTargetModel = createRadarTargetArrivalMock();
-    const radarTargetModelToggleHaloSpy = sinon.spy(radarTargetModel, 'toggleHalo');
-    const expectedResponse = [true, 'TOGGLE HALO'];
-    const response = model.toggleHalo(radarTargetModel);
+    const setHaloRadiusStub = sinon.stub(radarTargetModel, 'setHalo');
+    const expectedResponse = [false, 'ERR: HALO SIZE INVALID'];
+    const response = scopeModel.setHalo(radarTargetModel, 0);
 
     t.deepEqual(response, expectedResponse);
-    t.true(radarTargetModelToggleHaloSpy.called);
+    t.true(setHaloRadiusStub.notCalled);
+});
+
+ava('.setHalo() returns error when requested halo size is too large', (t) => {
+    const scopeModel = new ScopeModel();
+    const radarTargetModel = createRadarTargetArrivalMock();
+    const maxRadius = scopeModel._theme.SCOPE.HALO_MAX_RADIUS_NM;
+    const radius = maxRadius + 0.1;
+    const setHaloRadiusStub = sinon.stub(radarTargetModel, 'setHalo');
+    const expectedResponse = [false, `ERR: HALO MAX ${maxRadius} NM`];
+    const response = scopeModel.setHalo(radarTargetModel, radius);
+
+    t.deepEqual(response, expectedResponse);
+    t.true(setHaloRadiusStub.notCalled);
+});
+
+ava('.setHalo() uses default halo radius when one is not specified in the command', (t) => {
+    const scopeModel = new ScopeModel();
+    const radarTargetModel = createRadarTargetArrivalMock();
+    const defaultRadius = scopeModel._theme.SCOPE.HALO_DEFAULT_RADIUS_NM;
+    const setHaloRadiusStub = sinon.stub(radarTargetModel, 'setHalo');
+
+    scopeModel.setHalo(radarTargetModel);
+
+    t.true(setHaloRadiusStub.calledWithExactly(defaultRadius));
+});
+
+ava('.setHalo() calls RadarTargetModel.setHalo()', (t) => {
+    const scopeModel = new ScopeModel();
+    const radarTargetModel = createRadarTargetArrivalMock();
+    const radius = 7;
+    const setHaloRadiusStub = sinon.stub(radarTargetModel, 'setHalo');
+
+    scopeModel.setHalo(radarTargetModel, radius);
+
+    t.true(setHaloRadiusStub.calledWithExactly(radius));
 });
 
 ava('._setTheme returns early when an invalid theme name is passed', (t) => {
@@ -169,45 +214,5 @@ ava('._setTheme() changes the value of #_theme', (t) => {
 
     t.true(model._theme === THEME.CLASSIC);
 });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
-
-// // ava('', (t) => {
-//     //
-// });
 
 // ava.todo('Add test for .amendAltitude()');